tests/requirements.txt
python-logstash
elasticsearch